feat: support $NO_UPLOAD_MOD envvar
Checks $NO_UPLOAD_MOD envvar. If it is set and nonempty, all code
relating to uploading mods is disabled (including checks for keys). This
is helpful for contributors that do not have access to the upload keys
or for those who wish to build locally (e.g., for testing) without
uploading.